Initialized the UI
Figured out how to get the display working, as well as button interactions and text fields have basic test functions. Next up is adding memory storage for inputted text and then routing for the server. Then a new page will need to be created for interacting with the server properly.